Lattice MBR now uses nbest list as default hypothesis set during reranking. To use lattice instead as hypothesis set, run with -lattice-hypo-set option
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@2843 1f5c12ca-751b-0410-a591-d2e778427230

+++ moses-cmd/src/LatticeMBR.cpp
@@ -224,7 +224,7 @@ void calcNgramPosteriors(Lattice & connectedHyp, map<const Hypothesis*, vector<E
     }
     
     //Process ngrams now
-    for (int j =0 ; j < edges.size(); ++j) {
+    for (size_t j =0 ; j < edges.size(); ++j) {
       Edge& edge = edges[j];
       const NgramHistory & incomingPhrases = edge.GetNgrams(incomingEdges);
